Bug #1408487 cpufreq: sprd: fix sharkl3 dvfs PC-lint warning
[root cause  ] Pclint warning
[changes     ] fix Pclint warning
[side effects] none
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]:yes
[is there dependence]: none
[confirm dependent commit]: none
[board]: s9863a-1h10
[test case]: DVFS
[reviewers   ] luyao.wu, jingchao.ye, ling.xu
[change_type ] feature_bugfix
[tag_product ] common

Variable type conversion and return value type adjustment.

+++ drivers/cpufreq/sprd-hwdvfs-sharkl3.h
@@ -89,8 +89,8 @@
 
 /* REG_DVFS_CTRL_VERSION */
 
-#define BIT_DVFS_CTRL_PROJ_NAME(x)               (((x) & 0xFFFF) << 16)
-#define BIT_DVFS_CTRL_VERSION(x)                 (((x) & 0xFFFF))
+#define BIT_DVFS_CTRL_PROJ_NAME(x)       ((unsigned int)(((x) & 0xFFFF) << 16))
+#define BIT_DVFS_CTRL_VERSION(x)         (((x) & 0xFFFF))
 
 /* REG_DVFS_CTRL_USER_LOCK00 */
